Import Statements:
React: This imports the necessary modules and components from the React library.
Link: The Link component is used for navigation between different routes in the application.
MdOutlineMenu, MdClose, MdOutlineAddShoppingCart: These are icons imported from the react-icons/md library, which provides Material Design icons for use in React applications.
GlobalState: This imports the global state context from the GlobalState file. This context is typically used to manage and share global state across various components in the application.
axios: Axios is a popular HTTP client for making requests to the server. It allows the application to communicate with the backend server and retrieve data or perform actions.
Header Function Component:

This is a functional component named Header. It represents the header section of the application, which usually contains navigation links, branding, and other important elements.
Global State Context:

It uses the useContext hook provided by React to access the global state context. This allows the component to access global state variables and functions defined in the GlobalState context.
Destructuring State:

The useContext(GlobalState) hook returns the global state object, which contains various state variables and functions. Here, the component destructures the userAPI object from the global state to access specific state variables like isLogged and isAdmin.
Imported Axios:

Axios is used to make HTTP requests to the server. In this component, it may be used for actions like logging out a user (axios.get('/user/logout')), retrieving data from the server, or sending form submissions.
Overall, this component serves as the header for the application, importing necessary dependencies for handling navigation, user authentication, and server communication. It's a crucial part of the UI that provides users with essential functionalities and options for interacting with the application.