logoutUser Function:
This function handles the logout process when the user clicks on the logout link.
It is an asynchronous function that sends a request to the server to log out the user by making a GET request to '/user/logout' endpoint using Axios.
After logging out, it clears the local storage using localStorage.clear().
Finally, it sets the isAdmin and isLogged state variables to false using their respective setter functions (setIsAdmin and setIsLogged).