loggedRouter Function:
This function returns a JSX fragment representing the navigation links for the logged-in user.
It contains two list items, each containing a Link component.
The first Link points to the '/history' route and displays 'HISTORY' as the link text.
The second Link is responsible for the logout action. It points to the home route ('/') and specifies the logoutUser function to be called when clicked.
Both links have text color styles for normal and hover states.
These functions are responsible for rendering different sets of navigation links based on the user's authentication status and role. The logoutUser function ensures proper logout functionality, while adminRouter and loggedRouter functions handle navigation for admin and logged-in users, respectively.