onChangeInput function:
This function is an event handler for input change events. It's triggered when the user types something into an input field.
It takes an event object e as its parameter, which contains information about the event that occurred.
Inside the function, object destructuring is used to extract the name and value properties from e.target. e.target refers to the input element that triggered the event.
The name property corresponds to the name attribute of the input field, while the value property contains the new value entered by the user.
setUser({ ...user, [name]: value }): This line updates the user state by spreading the existing properties of the user object and then updating the property specified by the name variable with the new value.