loginSubmit function:
This function is an event handler for form submission events. It's triggered when the user submits the login form.
It takes an event object e as its parameter.
e.preventDefault(): This method prevents the default behavior of form submission, which is to reload the page.
Inside a try block, an asynchronous HTTP POST request is made to the server endpoint 'http://localhost:5000/user/login' using the Axios library.
The data sent in the request is an object { ...user }, which spreads the properties of the user object. This includes the email and password properties entered by the user.
If the request is successful, the user is redirected to the root URL '/' using window.location.href = "/".
If an error occurs (e.g., authentication failure), the error message from the server response is displayed using the alert function.
In summary, these functions handle user input changes and form submissions for a login form. They update the user state with the entered email and password values and send them to the server for authentication.