Function Component:
DetailProduct: This is a functional component responsible for rendering the details of a product.
Inside the component, the params variable is initialized using the useParams hook to access the parameters of the current route, which likely includes the ID of the product being viewed.
The state variable is initialized using the useContext hook to access the global state stored in the GlobalState context.
The products variable is extracted from the state.productsAPI.products array obtained from the global state. It likely contains an array of products fetched from an API.
The detailProduct state variable is initialized using the useState hook to manage the details of the product being viewed.
The useEffect hook is used to update the detailProduct state when the component mounts or when the params.id or products array changes. It searches for the product with the matching ID in the products array and updates the detailProduct state with the found product.
The component returns a JSX structure that displays the details of the product, including its image, title, ID, price, description, content, and number sold.
It also includes a "Buy Now" button wrapped inside a <Link> component, which navigates the user to the homepage when clicked. The button has gradient background colors and transitions applied for visual appeal.