Total Price and Total Items Calculation:
totalPrice: It calculates the total price of all items in the cart. It uses the reduce method to iterate over each product in the cart, multiplying its price by its quantity, and summing up the results.
totalItems: It calculates the total number of items in the cart. Similar to totalPrice, it uses the reduce method to sum up the quantities of all products in the cart.