Handle Increment Function:
handleIncrement(productId): This function is invoked when the user clicks to increment the quantity of a product in the cart. It creates a new cart array (newCart) by mapping over the existing cart items. For the product with the specified productId, it increments the quantity by 1. Finally, it updates the cart state with the modified cart array.